ganti integer jadi bigInteger di jumlah dan saldo

+++ database/migrations/2024_03_28_031447_create_saldo_akuns_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->integer('id_kode_perkiraan');
             $table->string('bulan')->nullable();
             $table->year('tahun');
-            $table->integer('saldo_debet');
-            $table->integer('saldo_kredit');
+            $table->bigInteger('saldo_debet');
+            $table->bigInteger('saldo_kredit');
             $table->integer('is_saldo_awal')->default(0);
             $table->timestamps();
         });
